remove abstract layer
