+++ include/webview_candidate_window.hpp
@@ -1,7 +1,5 @@
-#ifndef WEBVIEW_CANDIDATE_WINDOW_H
-#define WEBVIEW_CANDIDATE_WINDOW_H
+#pragma once
 
-#include "candidate_window.hpp"
 #include "utility.hpp"
 #ifdef __EMSCRIPTEN__
 #include <emscripten.h>
@@ -9,11 +7,66 @@
 #include "webview.h"
 #endif
 #include <cstdint>
+#include <functional>
 #include <iostream>
 #include <nlohmann/json.hpp>
 #include <sstream>
+#include <string>
+#include <vector>
 
 namespace candidate_window {
+enum layout_t {
+    horizontal,
+    vertical,
+};
+
+enum format_t {
+    Underline = (1 << 3),
+    Highlight = (1 << 4),
+    Bold = (1 << 6),
+    Strike = (1 << 7),
+    Italic = (1 << 8),
+};
+
+template <typename T> using formatted = std::vector<std::pair<T, int>>;
+
+enum theme_t { system = 0, light = 1, dark = 2 };
+
+enum writing_mode_t { horizontal_tb = 0, vertical_rl = 1, vertical_lr = 2 };
+
+enum scroll_state_t { none = 0, ready = 1, scrolling = 2 };
+
+enum scroll_key_action_t {
+    one = 1,
+    two = 2,
+    three = 3,
+    four = 4,
+    five = 5,
+    six = 6,
+    up = 10,
+    down = 11,
+    left = 12,
+    right = 13,
+    home = 14,
+    end = 15,
+    page_up = 16,
+    page_down = 17,
+    expand = 18,
+    collapse = 19,
+    commit = 20
+};
+
+struct CandidateAction {
+    int id;
+    std::string text;
+};
+
+struct Candidate {
+    std::string text;
+    std::string label;
+    std::string comment;
+    std::vector<CandidateAction> actions;
+};
 
 enum CustomAPI : uint64_t { kCurl = 1 };
 
@@ -22,27 +75,27 @@ extern std::unordered_map<std::string, std::function<std::string(std::string)>>
     handlers;
 #endif
 
-class WebviewCandidateWindow : public CandidateWindow {
+class WebviewCandidateWindow {
   public:
     WebviewCandidateWindow();
     ~WebviewCandidateWindow();
-    void set_layout(layout_t layout) override;
+    void set_layout(layout_t layout);
     void update_input_panel(const formatted<std::string> &preedit,
                             int preedit_cursor,
                             const formatted<std::string> &auxUp,
-                            const formatted<std::string> &auxDown) override;
+                            const formatted<std::string> &auxDown);
     void set_candidates(const std::vector<Candidate> &candidates,
                         int highlighted, scroll_state_t scroll_state,
-                        bool scroll_start, bool scroll_end) override;
-    void scroll_key_action(scroll_key_action_t action) override;
-    void answer_actions(const std::vector<CandidateAction> &actions) override;
-    void set_theme(theme_t theme) override;
-    void set_writing_mode(writing_mode_t mode) override;
-    void set_style(const void *style) override;
-    void set_native_blur(bool enabled) override;
-    void set_native_shadow(bool enabled) override;
-    void show(double x, double y) override;
-    void hide() override;
+                        bool scroll_start, bool scroll_end);
+    void scroll_key_action(scroll_key_action_t action);
+    void answer_actions(const std::vector<CandidateAction> &actions);
+    void set_theme(theme_t theme);
+    void set_writing_mode(writing_mode_t mode);
+    void set_style(const void *style);
+    void set_native_blur(bool enabled);
+    void set_native_shadow(bool enabled);
+    void show(double x, double y);
+    void hide();
 
     void update_accent_color();
     void set_accent_color();
@@ -53,6 +106,45 @@ class WebviewCandidateWindow : public CandidateWindow {
     void load_plugins(const std::vector<std::string> &names);
     void unload_plugins();
 #endif
+
+    void set_init_callback(std::function<void()> callback) {
+        init_callback = callback;
+    }
+
+    void set_select_callback(std::function<void(int index)> callback) {
+        select_callback = callback;
+    }
+
+    void set_highlight_callback(std::function<void(int index)> callback) {
+        highlight_callback = callback;
+    }
+
+    void set_cursor_text(const std::string &text) { cursor_text_ = text; }
+    void set_highlight_mark_text(const std::string &text) {
+        highlight_mark_text_ = text;
+    }
+
+    void set_page_callback(std::function<void(bool)> callback) {
+        page_callback = callback;
+    }
+
+    void set_scroll_callback(std::function<void(int, int)> callback) {
+        scroll_callback = callback;
+    }
+
+    void set_paging_buttons(bool pageable, bool has_prev, bool has_next) {
+        pageable_ = pageable;
+        has_prev_ = has_prev;
+        has_next_ = has_next;
+    }
+
+    void set_ask_actions_callback(std::function<void(int index)> callback) {
+        ask_actions_callback = callback;
+    }
+
+    void set_action_callback(std::function<void(int index, int id)> callback) {
+        action_callback = callback;
+    }
 
   private:
 #ifndef __EMSCRIPTEN__
@@ -73,6 +165,19 @@ class WebviewCandidateWindow : public CandidateWindow {
                         // webview
 
   private:
+    std::function<void()> init_callback = []() {};
+    std::function<void(int index)> select_callback = [](int) {};
+    std::function<void(int index)> highlight_callback = [](int) {};
+    std::function<void(bool next)> page_callback = [](bool) {};
+    std::function<void(int, int)> scroll_callback = [](int, int) {};
+    std::function<void(int index)> ask_actions_callback = [](int) {};
+    std::function<void(int index, int id)> action_callback = [](int, int) {};
+    std::string cursor_text_ = "";
+    std::string highlight_mark_text_ = "";
+    bool pageable_ = false;
+    bool has_prev_ = false;
+    bool has_next_ = false;
+
     /* Platform-specific interfaces (implemented in 'platform') */
     void *create_window();
     void set_transparent_background();
@@ -200,4 +305,3 @@ async_on_main(std::function<void()> functor) {
 #endif
 
 } // namespace candidate_window
-#endif
